requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv 
pymupdf
wikipedia
langchain-text-splitters
sentence_transformers
langchain_huggingface
langchain_chroma
langchain-openai
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
faiss-cpu
uuid
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
duckduckgo-search